backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
langgraph>=0.0.30
langchain>=0.1.10
langchain-openai>=0.0.2
langchain-community>=0.0.38
langfuse==2.20.0
httpx==0.26.0
python-multipart==0.0.6
pillow==10.2.0
requests==2.31.0
aiofiles==23.2.1